requirements.txt
streamlit
streamlit-autorefresh
requests
pandas
pytz
plotly
caldav
python-dotenv
vobject
gspread
oauth2client
feedparser
